shell: move ccnl commands from example to shell
